+++ test/integration/s3_file/default_spec.rb
@@ -9,3 +9,12 @@ end
 describe file('/tmp/a_file') do
   it { should be_file }
 end
+
+describe file('/tmp/file_with_group_by_name') do
+  it { should be_file }
+  its('group') { should eq 'testgroup' }
+end
+
+describe file('/tmp/file_with_group_by_gid') do
+  it { should be_file }
+end