+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in each version of the aws cookbook.
 
+## Unreleased
+
 ## 9.0.16 - *2023-07-10*
 
 ## 9.0.15 - *2023-05-17*