+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This file is used to list changes made in each version of the aws cookbook.
 
 ## Unreleased
 
-Allow group attribute of s3_file to be String or Integer
+Allow attributes group and owner of s3_file to be String or Integer
 
 ## 9.0.16 - *2023-07-10*